Disable Powerline Separators in vim-airline
I've decided that I don't like them. Especially because they look really
bad in iTerm with transparency turned on. The really standout from the
background color that they are supposed to blend (seamlessly) with. It
also causes some 'hidden' characters (where the foreground and
background color are set to the same color as the surrounding background
color) to become visible as well.

(Note: This probably has something to do with the fact that the patched
characters are actually in the foreground rather than the background, so
anything that attempts to modify the background color won't deal well
with this)